requirements.txt
requests
tabulate
streamlit
gspread
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pandas
